SITES-22965: Adds test coverage
