+++ bundles/core/src/test/java/com/adobe/cq/commerce/core/components/internal/models/v2/teaser/CommerceTeaserImplTest.java
@@ -20,6 +20,7 @@ import java.util.List;
 import org.apache.commons.httpclient.HttpStatus;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.osgi.services.HttpClientBuilderFactory;
+import org.apache.sling.api.resource.PersistenceException;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.api.resource.ValueMap;
@@ -74,7 +75,6 @@ public class CommerceTeaserImplTest {
     private static final String CATEGORY_PAGE = "/content/category-page";
     private static final String PAGE = "/content/pageA";
     private static final String TEASER = "/content/pageA/jcr:content/root/responsivegrid/commerceteaser2";
-    private static final String TEASER_WITH_URL_PATH_FOR_CATEGORY = "/content/pageA/jcr:content/root/responsivegrid/commerceteaserwithcategoryselectionasurlpath";
 
     private Resource commerceTeaserResource;
     private CommerceTeaser commerceTeaser;
@@ -97,23 +97,15 @@ public class CommerceTeaserImplTest {
         context.registerAdapter(Resource.class, GraphqlClient.class, (Function<Resource, GraphqlClient>) input -> input.getValueMap().get(
             "cq:graphqlClient", String.class) != null ? graphqlClient : null);
 
-        // TODO: CIF-2469
-        // With a newer version of OSGI mock we could re-inject the reference into the existing UrlProviderImpl
-        // context.registerInjectActivateService(new SpecificPageStrategy(), "generateSpecificPageUrls", true);
-        SpecificPageStrategy specificPageStrategy = context.getService(SpecificPageStrategy.class);
-        Whitebox.setInternalState(specificPageStrategy, "generateSpecificPageUrls", true);
-    }
-
-    private void setupTest(String componentPath) throws Exception {
         // Mock resource and resolver
-        commerceTeaserResource = spy(context.resourceResolver().getResource(componentPath));
+        commerceTeaserResource = spy(context.resourceResolver().getResource(TEASER));
         ResourceResolver resolver = spy(commerceTeaserResource.getResourceResolver());
         when(commerceTeaserResource.getResourceResolver()).thenReturn(resolver);
         when(commerceTeaserResource.getResourceSuperType()).thenReturn("core/wcm/components/teaser/v1/teaser");
         when(commerceTeaserResource.adaptTo(ComponentsConfiguration.class)).thenReturn(MOCK_CONFIGURATION_OBJECT);
 
         Page page = spy(context.currentPage(PAGE));
-        context.currentResource(componentPath);
+        context.currentResource(TEASER);
         context.currentResource(commerceTeaserResource);
         context.registerService(PathProcessor.class, new MockPathProcessor());
         Resource pageResource = spy(page.adaptTo(Resource.class));
@@ -126,15 +118,20 @@ public class CommerceTeaserImplTest {
         slingBindings.put(WCMBindingsConstants.NAME_CURRENT_PAGE, page);
         slingBindings.put(WCMBindingsConstants.NAME_PAGE_MANAGER, page.getPageManager());
         ComponentManager componentManager = commerceTeaserResource.getResourceResolver().adaptTo(ComponentManager.class);
-        slingBindings.put(WCMBindingsConstants.NAME_COMPONENT, componentManager.getComponent(componentPath));
+        slingBindings.put(WCMBindingsConstants.NAME_COMPONENT, componentManager.getComponent(TEASER));
         Style style = mock(Style.class);
         when(style.get(anyString(), anyInt())).then(i -> i.getArgumentAt(1, Object.class));
         slingBindings.put("currentStyle", style);
+
+        // TODO: CIF-2469
+        // With a newer version of OSGI mock we could re-inject the reference into the existing UrlProviderImpl
+        // context.registerInjectActivateService(new SpecificPageStrategy(), "generateSpecificPageUrls", true);
+        SpecificPageStrategy specificPageStrategy = context.getService(SpecificPageStrategy.class);
+        Whitebox.setInternalState(specificPageStrategy, "generateSpecificPageUrls", true);
     }
 
     @Test
-    public void verifyProperties() throws Exception {
-        setupTest(TEASER);
+    public void verifyProperties() {
         commerceTeaser = context.request().adaptTo(CommerceTeaserImpl.class);
         Assert.assertNotNull("The CommerceTeaser object is null", commerceTeaser);
 
@@ -149,8 +146,7 @@ public class CommerceTeaserImplTest {
     }
 
     @Test
-    public void verifyActions() throws Exception {
-        setupTest(TEASER);
+    public void verifyActions() {
         commerceTeaser = context.request().adaptTo(CommerceTeaserImpl.class);
         List<ListItem> actionItems = commerceTeaser.getActions();
 
@@ -185,8 +181,7 @@ public class CommerceTeaserImplTest {
     }
 
     @Test
-    public void verifyNoActionsConfigured() throws Exception {
-        setupTest(TEASER);
+    public void verifyNoActionsConfigured() throws PersistenceException {
         context.resourceResolver().delete(commerceTeaserResource.getChild("actions"));
         commerceTeaser = context.request().adaptTo(CommerceTeaserImpl.class);
 
@@ -195,45 +190,8 @@ public class CommerceTeaserImplTest {
     }
 
     @Test
-    public void verifyJsonExport() throws Exception {
-        setupTest(TEASER);
+    public void verifyJsonExport() {
         commerceTeaser = context.request().adaptTo(CommerceTeaserImpl.class);
         Utils.testJSONExport(commerceTeaser, "/exporter/commerce-teaser-v2.json");
-    }
-
-    @Test
-    public void verifyActionsWhenCategoryIdentifierSetAsUrlPath() throws Exception {
-        setupTest(TEASER_WITH_URL_PATH_FOR_CATEGORY);
-        commerceTeaser = context.request().adaptTo(CommerceTeaserImpl.class);
-        List<ListItem> actionItems = commerceTeaser.getActions();
-
-        Assert.assertTrue(commerceTeaser.isActionsEnabled());
-        Assert.assertTrue(actionItems.size() == 5);
-
-        // Product slug is configured and there is a dedicated specific subpage for that product
-        Assert.assertEquals(PRODUCT_SPECIFIC_PAGE + ".html/beaumont-summit-kit.html", actionItems.get(0).getURL());
-        Assert.assertEquals("A product", actionItems.get(0).getTitle());
-        Assert.assertEquals("MJ01", ((CommerceTeaserActionItem) actionItems
-            .get(0)).getEntityIdentifier().getValue());
-
-        // Category id is configured
-        Assert.assertEquals(CATEGORY_PAGE + ".html/equipment.html", actionItems.get(1).getURL());
-        Assert.assertEquals("A category", actionItems.get(1).getTitle());
-        Assert.assertEquals("equipment",
-            ((CommerceTeaserActionItem) actionItems.get(1)).getEntityIdentifier().getValue());
-
-        // Both are configured, category links "wins"
-        Assert.assertEquals(CATEGORY_PAGE + ".html/equipment.html", actionItems.get(2).getURL());
-        Assert.assertEquals("A category", actionItems.get(2).getTitle());
-        Assert.assertEquals("equipment", ((CommerceTeaserActionItem) actionItems.get(2))
-            .getEntityIdentifier().getValue());
-
-        // Some text is entered, current page is used
-        Assert.assertEquals(PAGE + ".html", actionItems.get(3).getURL());
-        Assert.assertEquals("Some text", actionItems.get(3).getTitle());
-
-        // Link is configured
-        Assert.assertEquals("/content/page.html", actionItems.get(4).getURL());
-        Assert.assertEquals("A page", actionItems.get(4).getTitle());
     }
 }
